+++ tedi/components/form/index.ts
@@ -5,3 +5,5 @@ export * from "./number-field/number-field.component";
 export * from "./select";
 export * from "./toggle/toggle.component";
 export * from "./date-picker/date-picker.component";
+export * from "./form-field/form-field.component";
+export * from "./text-field/text-field.component";